# docs/architecture/openapi.yaml
...
openapi: 3.1.0
info:
  title: xLearn BFF API
  version: 1.0.0
  description: |
    The only public xLearn API — the gateway's BFF, same-origin with the SPA at
    `projects.sujaykumar.dev/xlearn/api/v1` (cookie auth, no CORS). The unversioned
    `/xlearn/api` prefix is kept as a compat alias (ADR-0021).

    **Auth.** An HttpOnly `Secure` `SameSite=Lax` session cookie (`xl_session`). The
    SPA sends it with `credentials: include`. Unauthenticated requests get `401` with
    `{"error":{"code":"unauthenticated"}}`, on which the SPA redirects to OAuth. The
    gateway validates the session against identity and mints a short-TTL RS256 JWT per
    downstream audience (ADR-0006); clients never see it.

    **Errors.** Every error is the envelope `{"error":{"code","message"?,"details"?}}`;
    the HTTP status reflects the class (`400/401/403/404/409/422/429/5xx`). A failed
    upstream is `502 {"error":{"code":"upstream"}}`; an unconfigured one is `503`.

    **Aggregations.** Endpoints marked `agg` fan out to several services server-side
    (Week, Problem, Dashboard, Progress). Each degrades a section independently rather
    than failing the whole response, except `GET /progress`, whose summary leg is
    required.

    **Streaming.** `POST /coach/chat` responds `text/event-stream` (SSE): `data:` frames
    carrying `{"delta"}`, a terminal `event: done`, or `event: error` on a mid-stream
    failure. Errors before the first byte are the normal JSON envelope.
servers:
...
security:
  - cookieAuth: []
paths:
...
  /paths:
    get:
      tags: [curriculum]
      operationId: listPaths
      summary: Catalog — all learning paths + status.
      responses:
        "200":
          description: All paths.
          content:
            application/json:
              schema:
                type: object
                properties:
                  paths:
                    type: array
                    items: { $ref: "#/components/schemas/PathSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /paths/{slug}:
    get:
      tags: [curriculum]
      operationId: getPath
      summary: Roadmap — phases, weeks, totals.
      parameters:
        - { $ref: "#/components/parameters/Slug" }
      responses:
        "200":
          description: The path roadmap.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Roadmap" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
  /paths/{slug}/weeks/{n}:
    get:
      tags: [curriculum]
      operationId: getWeek
      summary: Week thesis + concepts + problems with the user's five-touch state (agg).
      parameters:
        - { $ref: "#/components/parameters/Slug" }
        - { name: n, in: path, required: true, schema: { type: integer, minimum: 1 } }
      responses:
        "200":
          description: Week content with the injected per-user `userState` block.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WeekView" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
  /concepts/{slug}:
    get:
      tags: [curriculum]
      operationId: getConcept
      summary: Concept reading + code template.
      parameters:
        - { $ref: "#/components/parameters/Slug" }
      responses:
        "200":
          description: The concept.
          content:
            application/json:
              schema:
                type: object
                properties:
                  concept: { $ref: "#/components/schemas/Concept" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /problems/{id}:
    get:
      tags: [problem]
      operationId: getProblem
      summary: Problem + only the unlocked stage sections + user state + timer (agg).
      parameters:
        - { $ref: "#/components/parameters/ProblemID" }
      responses:
        "200":
          description: Problem content filtered to unlocked stages, plus practice state.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemView" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: xl_session
      description: HttpOnly session cookie set by the OAuth callback.
  parameters:
    Slug:
      name: slug
      in: path
      required: true
      schema: { type: string }
    ProblemID:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: No/invalid session; the SPA redirects to OAuth.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example: { error: { code: unauthenticated } }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    BadGateway:
      description: An upstream service failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example: { error: { code: upstream, message: review unavailable } }
    Unavailable:
      description: A required upstream is not configured.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    Error:
      type: object
      required: [code]
      properties:
        code: { type: string }
        message: { type: string }
        details: {}
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { $ref: "#/components/schemas/Error" }
    Difficulty:
      type: string
      enum: [easy, med, hard]
    Outcome:
      type: string
      enum: [clean, rough, assisted, miss]
...
    PathSummary:
      type: object
      properties:
        slug: { type: string }
        title: { type: string }
        status: { type: string }
        summary: { type: string }
        problem_total: { type: integer }
        week_total: { type: integer }
    Phase:
      type: object
      properties:
        order: { type: integer }
        name: { type: string }
        theme: { type: string }
        week_from: { type: integer }
        week_to: { type: integer }
    WeekSummary:
      type: object
      properties:
        n: { type: integer }
        title: { type: string }
        thesis: { type: string }
        easy: { type: integer }
        med: { type: integer }
        hard: { type: integer }
        total: { type: integer }
    Roadmap:
      type: object
      properties:
        path: { $ref: "#/components/schemas/PathSummary" }
        phases: { type: array, items: { $ref: "#/components/schemas/Phase" } }
        weeks: { type: array, items: { $ref: "#/components/schemas/WeekSummary" } }
    Problem:
      type: object
      properties:
        id: { type: string }
        path_slug: { type: string }
        week_n: { type: integer }
        title: { type: string }
        difficulty: { $ref: "#/components/schemas/Difficulty" }
        pattern: { type: string }
        leetcode_url: { type: string }
        neetcode_url: { type: string }
        is_reinforcement: { type: boolean }
    Concept:
      type: object
      properties:
        slug: { type: string }
        path_slug: { type: string }
        title: { type: string }
        body_md: { type: string }
        when_to_use_md: { type: string }
        code_template: { type: string }
    Touch:
      type: object
      properties:
        level: { type: integer, description: "1..5 = Day 1/3/7/21/45" }
        dueDate: { type: [string, "null"] }
        result: { type: string, enum: [none, pass, fail] }
    ProblemState:
      type: object
      properties:
        status: { type: string, enum: [locked, available, attempting, solved] }
        lastOutcome:
          oneOf:
            - { $ref: "#/components/schemas/Outcome" }
            - { type: "null" }
        currentTouch: { type: integer }
        touches: { type: array, items: { $ref: "#/components/schemas/Touch" } }
    WeekView:
      type: object
      description: Curriculum week content plus the injected per-user `userState` (ADR-0013).
      properties:
        week: { type: object }
        path: { type: object }
        concepts: { type: array, items: { type: object } }
        problems: { type: array, items: { $ref: "#/components/schemas/Problem" } }
        phase: { type: object }
        userState:
          type: object
          properties:
            week:
              type: object
              properties:
                solved: { type: integer }
                coreTotal: { type: integer }
                byDifficulty:
                  type: object
                  properties:
                    easy: { type: integer }
                    med: { type: integer }
                    hard: { type: integer }
                populated: { type: boolean }
            problems:
              type: object
              additionalProperties: { $ref: "#/components/schemas/ProblemState" }
    Timer:
      type: object
      properties:
        kind: { type: string }
        deadlineAt: { type: string, format: date-time }
        remainingSeconds: { type: integer }
        expired: { type: boolean }
    PracticeState:
      type: object
      properties:
        problemId: { type: string }
        status: { type: string, enum: [locked, available, attempting, solved] }
        stageReached: { type: string }
        unlockedStages: { type: array, items: { type: string } }
        currentTouch: { type: integer }
        lastOutcome: { type: [string, "null"] }
        firstSolvedAt: { type: [string, "null"] }
        revealedEarly: { type: boolean }
        timer:
          oneOf:
            - { $ref: "#/components/schemas/Timer" }
            - { type: "null" }
    ProblemDetail:
      description: >-
        GET /problems/{id}'s problem: the Problem shape plus two additive, answer-free fields
        (m3-01) that list and bulk routes do not carry.
      allOf:
        - { $ref: "#/components/schemas/Problem" }
        - type: object
          properties:
            contract_hash_prefix:
              type: string
              description: First 12 hex chars of the item's contract_hash (humans see a prefix); empty on the self path.
            grading_summary:
              type: object
              description: How the item is graded, derived at seed. A self-path item is just {"mode":"self"}.
              properties:
                mode: { type: string, enum: [self, auto, mixed] }
                parts:
                  type: array
                  items:
                    type: object
                    properties:
                      id: { type: string }
                      type: { type: string }
                      grading: { type: string, enum: [auto, ai, none] }
                      cadence: { type: string, enum: [iterate, final] }
                grader_kinds: { type: array, items: { type: string } }
                languages: { type: array, items: { type: string } }
    ProblemView:
      type: object
      properties:
        problem: { $ref: "#/components/schemas/ProblemDetail" }
        sections:
          type: array
          description: Only the learner's UNLOCKED stage sections (R-PF1).
          items:
            type: object
            properties:
              stage: { type: string }
              kind: { type: string }
              order: { type: integer }
              body_md: { type: string }
              code: { type: string }
        state: { $ref: "#/components/schemas/PracticeState" }
...